requirements.txt
Django==1.6.5
dj-database-url==0.2.2
dj-static==0.0.5
django-secure==1.0
gunicorn==18.0
psycopg2==2.5.1
static==1.1.1
django-bootstrap3==4.11.0
UniversalClient==0.5.2
google-api-python-client==1.2
Pillow==2.5.0
python-dateutil==2.1
django-localflavor==1.0
South==0.8.2
stripe==1.11.0
six==1.7.3
pyrax==1.8.1 # 1.8.2 has bug that prevents cumulus working with ssl
django-cumulus==1.0.12
django-extensions==1.3.8
markdown==2.4.1
django-markupfield==1.2.0
django-pagedown==0.1.0
feedparser==5.1.3
